Introduce fetcher to separate fetching contributors from the presenter
