+++ workshops/AndroidKotlinWorkshop/app/src/main/java/com/novoda/workshop/contributors/Contributor.kt
@@ -3,7 +3,7 @@ package com.novoda.workshop.contributors
 import com.squareup.moshi.Json
 
 internal data class Contributor(
-    val name: String,
+    @field:Json(name = "login") val name: String,
     val contributions: Int,
     @field:Json(name = "avatar_url") val avatarUrl: String
 )
